error handling for scripts

+++ delete_all_saved_games.py
@@ -5,13 +5,20 @@ from constants import GAME_BACKUP_DIR
 from utils import delete_game_backup, get_yes_or_no_response, response_is_yes_and_not_empty
 
 if __name__ == '__main__':
-    print(f"Saved game files:\n{np.array(os.listdir(GAME_BACKUP_DIR))}")
-    prompt = f"Are you sure you want to delete all saved games? This will delete the files listed above. " \
-             f"(y/n) (default=n):"
-    response = get_yes_or_no_response(prompt)
-    if response_is_yes_and_not_empty(response):
-        for saved_file in os.listdir(GAME_BACKUP_DIR):
-            game_id = saved_file.split('.')[0]
-            delete_game_backup(game_id)
+    if os.path.exists(GAME_BACKUP_DIR):
+        print(f"Saved game files:\n{np.array(os.listdir(GAME_BACKUP_DIR))}")
+        prompt = f"Are you sure you want to delete all saved games? This will delete the files listed above. " \
+                 f"(y/n) (default=n):"
+        response = get_yes_or_no_response(prompt)
+        if response_is_yes_and_not_empty(response):
+            all_saved_files = os.listdir(GAME_BACKUP_DIR)
+            if len(all_saved_files) == 0:
+                print(f"No saved games. Save directory '{GAME_BACKUP_DIR}' is empty.")
+            else:
+                for saved_file in all_saved_files:
+                    game_id = saved_file.split('.')[0]
+                    delete_game_backup(game_id)
+        else:
+            print(f"Cancelling. No files deleted.")
     else:
-        print(f"Cancelling. No files deleted.")
+        print(f"No saved games. Save directory '{GAME_BACKUP_DIR}' does not exist.")
